// src/renderer/src/components/right-sidebar/checks-panel-git-effects-types.ts
export type ChecksPanelGitEffectsKey =
  | 'activeConnectionId'
  | 'activeWorktreeId'
  | 'activeWorktreePath'
  | 'activeWorktreePushTarget'
  | 'branch'
  | 'conflictSummaryRefreshKeyRef'
  | 'eligibilityHeadOidRef'
  | 'fallbackGitHubPRNumber'
  | 'fetchPRForBranch'
  | 'getHostedReviewCreationEligibility'
  | 'gitStatusInvalidation'
  | 'gitStatusReadyForPanelContext'
  | 'gitStatusSnapshotInFlightContextRef'
  | 'gitStatusSnapshotRerunContextRef'
  | 'gitStatusSnapshotRetryTimerRef'
  | 'hasUncommittedChanges'
  | 'hostedReviewCreationRequestKey'
  | 'isFolder'
  | 'isPanelVisible'
  | 'linkedAzureDevOpsPR'
  | 'linkedBitbucketPR'
  | 'linkedGitLabMR'
  | 'linkedGiteaPR'
  | 'linkedPR'
  | 'localExecutionScope'
  | 'ownerSettings'
  | 'panelContextKey'
  | 'panelContextKeyRef'
  | 'pr'
  | 'prCacheKey'
  | 'remoteStatus'
  | 'remoteStatusInvalidation'
  | 'repo'
  | 'repoConnectionId'
  | 'runtimeEnvironmentId'
  | 'setConflictDetailsRefreshing'
  | 'setGitStatusProbeErrorContextKey'
  | 'setGitStatusRefreshNonce'
  | 'setGitStatusSnapshot'
  | 'setHostedReviewCreationSnapshot'
  | 'settings'
  | 'sshConnectionStatus'
  | 'updateWorktreeGitIdentity'
;